Changed the way the manager instance is created
Each test suite run must have a single manager object. For regular runs,
a simple singleton would suffice, but for our own tests this causes
problems, as the nested pytest runs get the same instance. This can be
worked around by running them all in subprocesses, which we did until
now, but this slows the tests down massively. Instead, keep a stack of
managers, so each test suite will have its own manager, even nested
ones.

This is not foolproof, and definitely not compatible with xdist, but it
works well enough.